Move debounces and language extensions out of Codemirror.
